requirements.txt
opencv-python
numpy
torch
torchvision
matplotlib
scipy
Pillow
tensorboard
ipykernel
argparse
tqdm
scikit-learn
imageio
pandas
#tkinter
accelerate
diffusers
Flask
albumentations